refactor: int -> Long

+++ backend/src/main/java/com/codecool/manhwalabbackend/repository/GenreRepository.java
@@ -13,7 +13,7 @@ public interface GenreRepository extends JpaRepository<Genre, Long> {
     @Query(value = "SELECT genre.id, genre.name FROM genre " +
             "JOIN genre_for_comic gfc on genre.id = gfc.genre_id " +
             "WHERE comic_profile_id = ?1", nativeQuery = true)
-    List<Genre> getComicGenres(int comicId);
+    List<Genre> getComicGenres(Long comicId);
 
 }
